requirements.txt
# Core
ccxt>=4.2.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Async
asyncio-throttle>=1.0.2

# Utilities
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Dev / linting (optional)
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
